@@ -7,8 +7,8 @@
         "axios": "^0.21.1",
         "connected-react-router": "^6.9.1",
         "js-cookie": "^3.0.0",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-redux": "^7.2.4",
         "react-router-dom": "^5.2.0",
         "redux-saga": "^1.1.3"
